refactor(mcp): extract consent page template (kill-list B7)
